+++ android/src/main/java/com/crazecoder/flutterbugly/FlutterBuglyPlugin.java
@@ -13,10 +13,7 @@ import com.tencent.bugly.beta.UpgradeInfo;
 import com.tencent.bugly.beta.upgrade.UpgradeListener;
 import com.tencent.bugly.crashreport.CrashReport;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
+import java.util.Map;
 
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
@@ -100,10 +97,16 @@ public class FlutterBuglyPlugin implements MethodCallHandler {
                     if (!TextUtils.isEmpty(channel))
                         Bugly.setAppChannel(activity.getApplicationContext(), channel);
                 }
-                result(getResultBean(true, appId,"Bugly 初始化成功"));
+                result(getResultBean(true, appId, "Bugly 初始化成功"));
             } else {
-                result(getResultBean(false, null,"Bugly appId不能为空"));
+                result(getResultBean(false, null, "Bugly appId不能为空"));
             }
+        } else if (call.method.equals("setAppChannel")) {
+            if (call.hasArgument("channel")) {
+                String channel = call.argument("channel");
+                Bugly.setAppChannel(activity.getApplicationContext(), channel);
+            }
+            result(null);
         } else if (call.method.equals("setUserId")) {
             if (call.hasArgument("userId")) {
                 String userId = call.argument("userId");
@@ -141,12 +144,12 @@ public class FlutterBuglyPlugin implements MethodCallHandler {
             callback = new UpgradeCallback() {
                 @Override
                 public void onUpgrade(UpgradeInfo strategy) {
-                    if(finalUseCache){
+                    if (finalUseCache) {
                         if (strategy != null) {
                             upgradeInfo = strategy;
                         }
                         result(upgradeInfo);
-                    }else {
+                    } else {
                         result(strategy);
                     }
                 }
@@ -164,9 +167,11 @@ public class FlutterBuglyPlugin implements MethodCallHandler {
         }
 
     }
-    private void postException(MethodCall call){
+
+    private void postException(MethodCall call) {
         String message = "";
         String detail = null;
+        Map<String, String> map = null;
         if (call.hasArgument("crash_message")) {
             message = call.argument("crash_message");
         }
@@ -174,45 +179,50 @@ public class FlutterBuglyPlugin implements MethodCallHandler {
             detail = call.argument("crash_detail");
         }
         if (TextUtils.isEmpty(detail)) return;
-        String[] details = detail.split("#");
-        List<StackTraceElement> elements = new ArrayList<>();
-        for (String s : details) {
-            if (!TextUtils.isEmpty(s)) {
-                String methodName = null;
-                String fileName = null;
-                int lineNum = -1;
-                String[] contents = s.split(" \\(");
-                if (contents.length > 0) {
-                    methodName = contents[0];
-                    if (contents.length < 2) {
-                        break;
-                    }
-                    String packageContent = contents[1].replace(")", "");
-                    String[] packageContentArray = packageContent.split("\\.dart:");
-                    if (packageContentArray.length > 0) {
-                        if (packageContentArray.length == 1) {
-                            fileName = packageContentArray[0];
-                        } else {
-                            fileName = packageContentArray[0] + ".dart";
-                            Pattern patternTrace = Pattern.compile("[1-9]\\d*");
-                            Matcher m = patternTrace.matcher(packageContentArray[1]);
-                            if (m.find()) {
-                                String lineNumStr = m.group();
-                                lineNum = Integer.parseInt(lineNumStr);
-                            }
-                        }
-                    }
-                }
-                StackTraceElement element = new StackTraceElement("Dart", methodName, fileName, lineNum);
-                elements.add(element);
-            }
+        if (call.hasArgument("crash_data")) {
+            map = call.argument("crash_data");
         }
-        Throwable throwable = new Throwable(message);
-        if (elements.size() > 0) {
-            StackTraceElement[] elementsArray = new StackTraceElement[elements.size()];
-            throwable.setStackTrace(elements.toArray(elementsArray));
-        }
-        CrashReport.postCatchedException(throwable);
+        CrashReport.postException(8, "Flutter Exception", message, detail, map);
+
+//        String[] details = detail.split("#");
+//        List<StackTraceElement> elements = new ArrayList<>();
+//        for (String s : details) {
+//            if (!TextUtils.isEmpty(s)) {
+//                String methodName = null;
+//                String fileName = null;
+//                int lineNum = -1;
+//                String[] contents = s.split(" \\(");
+//                if (contents.length > 0) {
+//                    methodName = contents[0];
+//                    if (contents.length < 2) {
+//                        break;
+//                    }
+//                    String packageContent = contents[1].replace(")", "");
+//                    String[] packageContentArray = packageContent.split("\\.dart:");
+//                    if (packageContentArray.length > 0) {
+//                        if (packageContentArray.length == 1) {
+//                            fileName = packageContentArray[0];
+//                        } else {
+//                            fileName = packageContentArray[0] + ".dart";
+//                            Pattern patternTrace = Pattern.compile("[1-9]\\d*");
+//                            Matcher m = patternTrace.matcher(packageContentArray[1]);
+//                            if (m.find()) {
+//                                String lineNumStr = m.group();
+//                                lineNum = Integer.parseInt(lineNumStr);
+//                            }
+//                        }
+//                    }
+//                }
+//                StackTraceElement element = new StackTraceElement("Dart", methodName, fileName, lineNum);
+//                elements.add(element);
+//            }
+//        }
+//        Throwable throwable = new Throwable(message);
+//        if (elements.size() > 0) {
+//            StackTraceElement[] elementsArray = new StackTraceElement[elements.size()];
+//            throwable.setStackTrace(elements.toArray(elementsArray));
+//        }
+//        CrashReport.postCatchedException(throwable);
     }
 
     private void result(Object object) {
@@ -226,7 +236,7 @@ public class FlutterBuglyPlugin implements MethodCallHandler {
         }
     }
 
-    private BuglyInitResultInfo getResultBean(boolean isSuccess,String appId, String msg) {
+    private BuglyInitResultInfo getResultBean(boolean isSuccess, String appId, String msg) {
         BuglyInitResultInfo bean = new BuglyInitResultInfo();
         bean.setSuccess(isSuccess);
         bean.setAppId(appId);