+++ ios/Classes/FlutterBuglyPlugin.m
@@ -46,10 +46,13 @@
       if ([crash_detail isKindOfClass:[NSNull class]]) {
           crash_message = @"";
       }
-      NSException* ex = [[NSException alloc]initWithName:crash_message
-                                                  reason:crash_detail
-                                                userInfo:nil];
-      [Bugly reportException:ex];
+      NSArray *stackTraceArray = [crash_detail componentsSeparatedByString:@""];
+      NSDictionary *data = call.arguments[@"crash_data"];
+      if(data == nil){
+        data = [NSMutableDictionary dictionary];
+      }
+
+      [Bugly reportExceptionWithCategory:5 name:crash_message reason:@" " callStack:stackTraceArray extraInfo:data terminateApp:NO];
       result(nil);
   }else if([@"setUserId" isEqualToString:call.method]){
       NSString *userId = call.arguments[@"userId"];